requirements.txt
autopep8==1.6.0
requests==2.27.1
python-dotenv==0.19.2